refactor: move debounce to its own file
